% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-power-by-n.R
\name{plotPowerByN}
\alias{plotPowerByN}
\title{Plot Power by Sample Size}
\usage{
plotPowerByN(V, Delta, min_Delta, alpha = 0.05, sample_size_grid,
  color = "black")
}
\arguments{
\item{V}{The covariance matrix of mean EDTR estimators.}

\item{Delta}{The vector of effect sizes with a zero indicating the best EDTR.}

\item{min_Delta}{The minimum desired detectable effect size.}

\item{alpha}{The Type I error rate for not including the true best EDTR.}

\item{sample_size_grid}{A vector of sample sizes.}

\item{color}{The color of the graph.}
}
\description{
Plots the power over a grid of sample sizes.
}
\details{
It employs common random variables
to reduce the variance. See \code{\link{computePower}} for more details.
}
