% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartsizer.R
\docType{package}
\name{smartsizer}
\alias{smartsizer}
\alias{smartsizer-package}
\title{smartsizer: A package for Sizing SMART Designs}
\description{
The smartsizer package is designed to assist investigators with sizing sequential,
multiple assignment, randomized trial (SMART) for determination of the optimal
dynamic treatment regime (DTR). smartsizer includes functions which permit
calculation of the minimum number of individuals to enroll in a SMART
in order to be able to detect a specified effect size between the best and
inferior embedded DTR, with a specified power.
smartsizer is designed for an arbitrary SMART design.
}
